requirements.txt
websockets==12.0
discord.py==2.3.2